Move tests into test_views, add test coverage for views in workup.
